fix(storage): close AN2 QA findings
